REFACTOR: Friends UI& 구조 변경
