+++ public/data/FamousFriends.json
@@ -4,33 +4,33 @@
       "Userid": "1",
       "name": "방탄소년단",
       "nickname": "유명인1",
-      "profile": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
+      "profileImage": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
       "famous": true,
-      "friends": true
+      "isfriend": true
     },
     {
       "Userid": "2",
       "name": "아이유",
       "nickname": "유명인2",
-      "profile": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
+      "profileImage": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
       "famous": true,
-      "friends": false
+      "isfriend": false
     },
     {
       "Userid": "3",
       "name": "뉴진스",
       "nickname": "유명인3",
-      "profile": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
+      "profileImage": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
       "famous": true,
-      "friends":false
+      "isfriend":false
     },
     {
       "Userid": "4",
       "name": "마동석",
       "nickname": "유명인4",
-      "profile": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
+      "profileImage": "https://w.namu.la/s/b08f047a5e40252885eade1e88cea066ff8fb41c304c167ae9ebe5677904fb907019b47eee3ac8a76cc05c2aec8353e64a3784b9f932b1309f395ba463e3b6292d435af094d4e4f4fd7a25baf0fc22e98da19b8f4dca8bee5a9d5486a703ea06e3eae74367aabce1866d6729b1a4b355",
       "famous": true,
-      "friends":false
+      "isfriend":false
     }
   ]
 }
