+++ src/components/friends/sidebar/FriendsSidebar.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useRef } from "react";
 import FriendList from "./FriendList";
 import axios from "axios";
 import { Link } from "react-router-dom";
